:recycle: Rename takeSnapshot to takeArchive.
This was previously aliased instead of being actually renamed. This renames
both the file in which the function is declared, as well as the function
itself.
