+++ src/index.ts
@@ -5,4 +5,4 @@ import { makeTest } from './playwright-api/makeTest';
 export const test = makeTest(base);
 export { expect };
 
-export { takeSnapshot as takeArchive } from './playwright-api/takeSnapshot';
+export { takeArchive } from './playwright-api/takeArchive';
